knowledge_base_qa/requirements.txt
langchain
langchain-community
langchain-core
redis>=5.0.0
huggingface-hub
sentence-transformers
pypdf
unstructured
python-dotenv
